Fixed missing renderIcon function
