+++ components/btngroup/BtnGroup.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import getItemKey from '../../utils/misc';
+import { getItemKey } from '../../utils/misc';
 import styles from './BtnGroup.module.scss';
 import BtnLink from '@/btnlink/BtnLink';
