+++ utils/renderContent.js
@@ -6,20 +6,20 @@
 
 // Dynamic utility that loads components.
 import dynamic from 'next/dynamic';
-import getitemKey from './misc';
+import { getItemKey } from './misc';
 
 export default function renderContent(item) {
 	switch (item.__component) {
 		// Title component.
 		case 'typography.title':
 			const Title = dynamic(() => import('@/title/Title'));
-			return <Title props={item} key={getitemKey(item.__component, item.id)} />;
+			return <Title props={item} key={getItemKey(item.__component, item.id)} />;
 
 		// Card component.
 		case 'layout.card':
 			const Card = dynamic(() => import('@/card/Card'));
 			return (
-				<Card key={getitemKey(item.__component, item.id)}>{item.content}</Card>
+				<Card key={getItemKey(item.__component, item.id)}>{item.content}</Card>
 			);
 
 		// Centered Content.
@@ -28,7 +28,7 @@ export default function renderContent(item) {
 			return (
 				<div
 					className='centeredContent'
-					key={getitemKey(item.__component, item.id)}
+					key={getItemKey(item.__component, item.id)}
 				>
 					<ReactMarkdown source={item.content} />
 				</div>
@@ -40,11 +40,11 @@ export default function renderContent(item) {
 			return (
 				<SplitContent
 					props={item}
-					key={getitemKey(item.__component, item.id)}
+					key={getItemKey(item.__component, item.id)}
 				/>
 			);
 
 		default:
-			return <p key={getitemKey(item.style, item.id)}>Component not found</p>;
+			return <p key={getItemKey(item.style, item.id)}>Component not found</p>;
 	}
 }
